Add PostgreSQL integration and REST API endpoints for cart and orders

+++ src/main/java/com/example/firstproject/UserRepository.java
@@ -2,7 +2,7 @@ package com.example.firstproject;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<Users,Integer>
+public interface UserRepository extends JpaRepository<Users,Long>
 {
    Users findByUsername(String n);
    Users findByEmail(String email);
